requirements.txt
alembic==1.12.1
bcrypt==4.0.1
black==23.10.1
blinker==1.6.3
click==8.1.7
exceptiongroup==1.1.3
Flask==3.0.0
Flask-Bcrypt==1.0.1
Flask-JWT-Extended==4.5.3
Flask-SQLAlchemy==3.1.1
greenlet==3.0.1
gunicorn==21.2.0
iniconfig==2.0.0
isort==5.12.0
itsdangerous==2.1.2
Jinja2==3.1.2
Mako==1.2.4
MarkupSafe==2.1.3
mypy-extensions==1.0.0
packaging==23.2
pathspec==0.11.2
platformdirs==3.11.0
pluggy==1.3.0
psycopg2-binary==2.9.9
PyJWT==2.8.0
pyscopg2==66.0.2
pytest==7.4.3
python-dotenv==1.0.0
SQLAlchemy==2.0.22
tomli==2.0.1
typing_extensions==4.8.0
Werkzeug==3.0.1
